+++ bootique-jdbc-junit5/src/test/resources/io/bootique/jdbc/junit5/changelogs/derby-changelog.sql
@@ -1,0 +1,13 @@
+--liquibase formatted sql
+
+--changeset derby:create-schema
+create table "b" (
+    "id" integer not null primary key,
+    "name" varchar(20)
+);
+
+--changeset derby:insert_data
+insert into "b" ("id", "name") values (
+  10, 'myname'
+);
+