+++ bootique-jdbc-junit5/src/test/resources/io/bootique/jdbc/junit5/changelogs/postgres-changelog.sql
@@ -1,0 +1,13 @@
+--liquibase formatted sql
+
+--changeset postgres:create-schema
+create table b (
+    id integer not null primary key,
+    name varchar(20)
+);
+
+--changeset postgres:insert_data
+insert into b values (
+  12, 'myname'
+);
+